Fix scala3 compile errors
It didn't compile with 3.0.1 due to one implicit not found problem.
But luckily, with 3.0.2 it works :-)

+++ modules/minio/src/main/scala/binny/minio/Minio.scala
@@ -80,7 +80,7 @@ final private[minio] class Minio[F[_]: Sync](client: MinioClient) {
     val args = range match {
       case ByteRange.All => aargs.build()
       case ByteRange.Chunk(offset, length) =>
-        aargs.offset(offset).length(length).build()
+        aargs.offset(offset).length(length.toLong).build()
     }
     Sync[F].blocking(client.getObject(args))
   }
